payment_gateway_ui/pyproject.toml
[project]
name = "bscore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
